+++ build-installer.ps1
@@ -1,3 +1,4 @@
+
 # build-installer.ps1
 # Build wtmux MSI installer using WiX Toolset
 #